[veomni] move the shard-export mechanics to utils.py (review)
transformer_impl keeps a four-line method; the DTensor + EP declaration
machinery (expert slot table, ep-param detection, the export generator) joins
the converter machinery already living in veomni/utils.py.
